refactor: move candidate preparation into framework
