+++ gateway-framework/src/gateway/http/mod.rs
@@ -1,2 +1,2 @@
+pub mod gateway;
 pub mod requests;
-pub use requests::{RequestSelector, RequestSelectorRejection};
