+++ gateway-framework/src/gateway/http/requests/mod.rs
@@ -1,6 +1,5 @@
 pub mod auth;
 pub mod blocks;
 pub mod budget;
-mod selector;
-
-pub use selector::{RequestSelector, RequestSelectorRejection};
+pub mod candidates;
+pub mod selector;
